pyproject.toml
[project]
name = "cc-boss"
version = "0.1.0"
description = "Orchestrate multiple Claude Code instances in parallel via web UI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "aiosqlite>=0.21",
    "pydantic>=2.10",
    "click>=8.1",
    "websockets>=14.0",
    "pyyaml>=6.0",
]

[project.scripts]
cc-boss = "cc_boss.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
